pyproject.toml
[build-system]
requires = ["setuptools>=57.6.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "speedwagon-uiucprescon"
version = "0.0.1.dev0"
dependencies = [
    "speedwagon==0.4.0b5",
    "HathiZip>=0.1.10",
    "HathiValidate>=0.3.8",
    "pyhathiprep>=0.1.10",
    "uiucprescon.imagevalidate>=0.1.9",
    "uiucprescon.ocr>=0.1.4",
    "uiucprescon.packager[kdu]>=0.2.15",
    "uiucprescon.images",
    "py3exiv2bind>=0.1.11",
    "pykdu-compress>=0.1.9",
    "typing_extensions;python_version<'3.11'"
]
license = { file="LICENSE"}
description = "Collection of tools and workflows for DS"
maintainers = [
    { name="Henry Borchers", email="hborcher@illinois.edu" }
]
authors = [
    {name="University Library at The University of Illinois at Urbana Champaign: Preservation Services", email="prescons@library.illinois.edu"}
]

[project.entry-points.'speedwagon.plugins']
#uiucprescon_workflows = 'speedwagon_uiucprecson:uiucprescon_plugin'
uiucprescon_active_workflows = 'speedwagon_uiucprescon.active_workflows'
uiucprescon_deprecated_workflows = 'speedwagon_uiucprescon.deprecated_workflows'
toy_workflows = 'speedwagon_uiucprescon.toy_workflows'

#uiucprescon_plugin_deprecated = 'speedwagon_uiucprecson:uiucprescon_plugin_deprecated'
[tool.setuptools]
packages = [
    'speedwagon_uiucprescon',
    'speedwagon_uiucprescon.tasks'
]

[tool.coverage.report]
exclude_lines = [
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
]
[tool.pylint.main]
extension-pkg-allow-list = [
    "py3exiv2bind","uiucprescon.ocr.tesseractwrap"
]

[tool.pylint."messages control"]
disable = [
    "raw-checker-failed",
    "bad-inline-option",
    "locally-disabled",
    "file-ignored",
    "suppressed-message",
    "useless-suppression",
    "deprecated-pragma",
    "use-symbolic-message-instead",
    "use-implicit-booleaness-not-comparison-to-string",
    "use-implicit-booleaness-not-comparison-to-zero",
#    Documentation checks should be done by pydocstyle instead of pylint
    "missing-docstring"
]
[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = ["E", "F", "W", "D", "B", "C"]

[tool.ruff.lint.pydocstyle]
convention = "google"
#
[tool.windows_standalone_packager.cpack_config_variables]
CPACK_WIX_UPGRADE_GUID = "BD289D57-8F94-4F4D-9B6D-82A532F3514C"
CPACK_PACKAGE_NAME = "Speedwagon (UIUC Prescon Edition)"
CPACK_PACKAGE_INSTALL_DIRECTORY = "Speedwagon - UIUC"
CPACK_PACKAGE_VENDOR = "University Library at The University of Illinois at Urbana Champaign: Preservation Services"
CPACK_WIX_PRODUCT_GUID = "84BB7CAB-24DA-46EC-A855-B972384534DB"